[B] Reorganized packages
- added model, persistence, utils and exception packages
- removed repository and dto package
- moved java classes accordingly
