+++ src/main/java/com/shop/restaurant/model/BoughtPositionWriteModel.java
@@ -1,4 +1,4 @@
-package com.shop.restaurant.model.dto;
+package com.shop.restaurant.model;
 
 public class BoughtPositionWriteModel {
