+++ src/main/java/com/shop/restaurant/model/OrderCommitWriteModel.java
@@ -1,4 +1,4 @@
-package com.shop.restaurant.model.dto;
+package com.shop.restaurant.model;
 
 public class OrderCommitWriteModel {
   private String comment;
